requirements.txt
# Core AI Framework
pydantic-ai>=0.0.14
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM Providers
openai>=1.0.0

# Database
asyncpg>=0.29.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# HTTP Client
aiohttp>=3.9.0

# Phase 2: Knowledge Graph & Neo4j
neo4j>=5.14.0
graphiti-core>=0.3.0

# Configuration
python-dotenv>=1.0.0

# Testing (optional but recommended)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Supabase (for AI PM Agent)
supabase>=2.0.0

# Additional utilities
aiofiles>=23.0.0
httpx>=0.25.0
loguru>=0.7.0
black>=23.0.0
ruff>=0.1.0
